Move session zarr checksum classes
