+++ dandiapi/zarr/tasks/__init__.py
@@ -1,6 +1,5 @@
 from __future__ import annotations
 
-from datetime import datetime
 from pathlib import Path
 
 from celery import shared_task
@@ -12,13 +11,7 @@ from django.db.transaction import atomic
 
 from dandiapi.api.asset_paths import add_zarr_paths, delete_zarr_paths
 from dandiapi.api.storage import get_boto_client, yield_files
-from dandiapi.zarr.checksums import (
-    ZarrChecksum,
-    ZarrChecksumFileUpdater,
-    ZarrChecksumListing,
-    ZarrChecksumModification,
-    ZarrChecksumUpdater,
-)
+from dandiapi.zarr.checksums import SessionZarrChecksumUpdater, ZarrChecksum
 from dandiapi.zarr.models import ZarrArchive, ZarrArchiveStatus
 
 logger = get_task_logger(__name__)
@@ -41,59 +34,6 @@ def clear_checksum_files(zarr: ZarrArchive):
         allowed = {'Key', 'VersionId'}
         objects = [{k: v for k, v in file.items() if k in allowed} for file in files]
         client.delete_objects(Bucket=bucket, Delete={'Objects': objects})
-
-
-class SessionZarrChecksumFileUpdater(ZarrChecksumFileUpdater):
-    """Override ZarrChecksumFileUpdater to ignore old files."""
-
-    def __init__(self, *args, session_start: datetime, **kwargs):
-        super().__init__(*args, **kwargs)
-
-        # Set the start of when new files may have been written
-        self._session_start = session_start
-
-    def read_checksum_file(self) -> ZarrChecksumListing | None:
-        """Load a checksum listing from the checksum file."""
-        storage = self.zarr_archive.storage
-        checksum_path = self.checksum_file_path
-
-        # Check if this file was modified during this session
-        read_existing = False
-        if storage.exists(checksum_path):
-            read_existing = storage.modified_time(self.checksum_file_path) >= self._session_start
-
-        # Read existing file if necessary
-        if read_existing:
-            with storage.open(checksum_path) as f:
-                x = f.read()
-                return self._serializer.deserialize(x)
-        else:
-            return None
-
-
-class SessionZarrChecksumUpdater(ZarrChecksumUpdater):
-    """ZarrChecksumUpdater to distinguish existing and new checksum files."""
-
-    def __init__(self, *args, **kwargs) -> None:
-        super().__init__(*args, **kwargs)
-
-        # Set microseconds to zero, since the timestamp from S3/Minio won't include them
-        # Failure to set this to zero will result in false negatives
-        self._session_start = datetime.now().replace(tzinfo=None, microsecond=0)
-
-    def apply_modification(
-        self, modification: ZarrChecksumModification
-    ) -> SessionZarrChecksumFileUpdater:
-        with SessionZarrChecksumFileUpdater(
-            zarr_archive=self.zarr_archive,
-            zarr_directory_path=modification.path,
-            session_start=self._session_start,
-        ) as file_updater:
-            file_updater.add_file_checksums(modification.files_to_update)
-            file_updater.add_directory_checksums(modification.directories_to_update)
-            file_updater.remove_checksums(modification.paths_to_remove)
-
-        return file_updater
 
 
 @shared_task(queue='ingest_zarr_archive', time_limit=3600)
